consolidate and standardize table typing
