+++ packages/Table/src/index.ts
@@ -3,7 +3,7 @@ import "./styles/global.css";
 import Table from "./Table";
 
 export type { TableProps } from "./Table";
-export type { TableConfig } from "./TableProperties";
+export type { TableOptions as TableConfig } from "./types";
 export type { TableColumn as TableColumn } from "./Column";
 export type { RowSelectionState } from "@tanstack/react-table";
 export * from "./ControlElements";
